refactor(redux): use ducks-like architecture
refactor(Root): take routes as prop rather than importing

refactor(webpack): use resolve.root rather than explicit aliases

refactor(DevTools): move from redux/middleware -> containers

docs(readme): update project structure

fix(createDevToolsWindow): cleanup bad imports
