--- src/actions/counter.js
@@ -1,5 +1,0 @@
-import { COUNTER_INCREMENT } from 'constants/counter';
-
-export default {
-  increment: () => ({ type : COUNTER_INCREMENT })
-};
