--- src/constants/counter.js
@@ -1,1 +1,0 @@
-export const COUNTER_INCREMENT = 'COUNTER_INCREMENT';
