+++ src/containers/Root.js
@@ -1,42 +1,37 @@
-import React                    from 'react';
-import { Provider }             from 'react-redux';
-import { Router }               from 'react-router';
-import routes                   from '../routes';
-import DevTools                 from './DevTools';
-import { createDevToolsWindow } from 'utils';
+import React        from 'react';
+import { Provider } from 'react-redux';
+import { Router }   from 'react-router';
+import DevTools     from 'containers/DevTools';
 
 export default class Root extends React.Component {
   static propTypes = {
     history : React.PropTypes.object.isRequired,
-    store   : React.PropTypes.object.isRequired,
-    debug   : React.PropTypes.bool,
-    debugExternal : React.PropTypes.bool
-  }
-
-  static defaultProps = {
-    debug : false,
-    debugExternal : false
-  }
-
-  renderDevTools () {
-    if (!this.props.debug) {
-      return null;
-    }
-
-    return this.props.debugExternal ?
-      createDevToolsWindow(this.props.store) : <DevTools />;
+    routes  : React.PropTypes.element.isRequired,
+    store   : React.PropTypes.object.isRequired
   }
 
   render () {
-    return (
-      <Provider store={this.props.store}>
-        <div>
-          <Router history={this.props.history}>
-            {routes}
-          </Router>
-          {this.renderDevTools()}
-        </div>
-      </Provider>
+    const content = (
+      <Router history={this.props.history}>
+        {this.props.routes}
+      </Router>
     );
+
+    if (__DEBUG__ && !__DEBUG_NW__) {
+      return (
+        <Provider store={this.props.store}>
+          <div>
+            {content}
+            <DevTools />
+          </div>
+        </Provider>
+      );
+    } else {
+      return (
+        <Provider store={this.props.store}>
+          {content}
+        </Provider>
+      );
+    }
   }
 }
