--- src/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers }    from 'redux';
-import { routeReducer }       from 'redux-simple-router';
-import counter                from './counter';
-
-export default combineReducers({
-  counter,
-  routing: routeReducer
-});
