+++ src/utils/createDevToolsWindow.js
@@ -3,7 +3,7 @@ import ReactDOM     from 'react-dom';
 import { Provider } from 'react-redux';
 import DevTools     from 'containers/DevToolsWindow';
 
-export function createDevToolsWindow (store) {
+export default function createDevToolsWindow (store) {
   const win = window.open(
     null,
     'redux-devtools', // give it a name so it reuses the same window
@@ -28,5 +28,3 @@ export function createDevToolsWindow (store) {
     );
   }, 10);
 }
-
-export default createDevToolsWindow;
