+++ src/views/HomeView.js
@@ -1,8 +1,8 @@
-import React          from 'react';
-import { connect }    from 'react-redux';
-import counterActions from 'actions/counter';
-import { Link }       from 'react-router';
-import styles         from './HomeView.scss';
+import React from 'react';
+import { connect } from 'react-redux';
+import { Link } from 'react-router';
+import { actions as counterActions } from '../redux/modules/counter';
+import styles from './HomeView.scss';
 
 // We define mapStateToProps where we'd normally use
 // the @connect decorator so the data requirements are clear upfront, but then
